细节调整
