+++ app/src/main/java/cn/wthee/pcrtool/database/AppPvpDatabase.kt
@@ -4,8 +4,6 @@ import android.annotation.SuppressLint
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import androidx.room.migration.Migration
-import androidx.sqlite.db.SupportSQLiteDatabase
 import cn.wthee.pcrtool.MyApplication
 import cn.wthee.pcrtool.data.db.dao.PvpDao
 import cn.wthee.pcrtool.data.db.entity.PvpFavoriteData
@@ -17,7 +15,7 @@ import cn.wthee.pcrtool.utils.Constants
         PvpFavoriteData::class,
         PvpHistoryData::class,
     ],
-    version = 101,
+    version = 102,
     exportSchema = false,
 )
 /**
