+++ app/src/main/java/cn/wthee/pcrtool/ui/skill/SkillSummonDetail.kt
@@ -14,7 +14,7 @@ import cn.wthee.pcrtool.data.model.CharacterProperty
 import cn.wthee.pcrtool.ui.MainActivity.Companion.navViewModel
 import cn.wthee.pcrtool.ui.common.*
 import cn.wthee.pcrtool.ui.theme.Dimen
-import cn.wthee.pcrtool.ui.tool.BossSkillList
+import cn.wthee.pcrtool.ui.tool.clan.BossSkillList
 import cn.wthee.pcrtool.utils.int
 import cn.wthee.pcrtool.viewmodel.CharacterAttrViewModel
 import cn.wthee.pcrtool.viewmodel.ClanViewModel
