+++ app/src/main/java/cn/wthee/pcrtool/ui/tool/CalendarEventList.kt
@@ -29,7 +29,7 @@ import kotlinx.coroutines.launch
 
 
 /**
- * 日程
+ * 日程记录
  */
 @Composable
 fun CalendarEventList(
