+++ app/src/main/java/cn/wthee/pcrtool/ui/tool/clan/ClanBattleDetail.kt
@@ -1,209 +1,47 @@
-package cn.wthee.pcrtool.ui.tool
+package cn.wthee.pcrtool.ui.tool.clan
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyGridState
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import cn.wthee.pcrtool.R
-import cn.wthee.pcrtool.data.db.view.ClanBattleInfo
 import cn.wthee.pcrtool.data.db.view.EnemyParameterPro
 import cn.wthee.pcrtool.data.enums.MainIconType
 import cn.wthee.pcrtool.data.enums.UnitType
-import cn.wthee.pcrtool.ui.PreviewBox
 import cn.wthee.pcrtool.ui.common.*
 import cn.wthee.pcrtool.ui.skill.SkillItem
 import cn.wthee.pcrtool.ui.skill.SkillLoopList
 import cn.wthee.pcrtool.ui.theme.CardTopShape
 import cn.wthee.pcrtool.ui.theme.Dimen
-import cn.wthee.pcrtool.ui.theme.FadeAnimation
 import cn.wthee.pcrtool.utils.ImageResourceHelper
-import cn.wthee.pcrtool.utils.ImageResourceHelper.Companion.ICON_UNIT
 import cn.wthee.pcrtool.utils.getZhNumberText
-import cn.wthee.pcrtool.utils.intArrayList
 import cn.wthee.pcrtool.viewmodel.ClanViewModel
 import cn.wthee.pcrtool.viewmodel.SkillViewModel
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.rememberPagerState
-import com.google.accompanist.placeholder.PlaceholderHighlight
-import com.google.accompanist.placeholder.material.placeholder
-import com.google.accompanist.placeholder.material.shimmer
-import kotlinx.coroutines.launch
 
-/**
- * 每月 BOSS 信息列表
- */
-@Composable
-fun ClanBattleList(
-    scrollState: LazyGridState,
-    toClanBossInfo: (Int, Int, Int) -> Unit,
-    clanViewModel: ClanViewModel = hiltViewModel()
-) {
-
-    val clanList =
-        clanViewModel.getAllClanBattleData().collectAsState(initial = arrayListOf()).value
-    val coroutineScope = rememberCoroutineScope()
-
-    Box(modifier = Modifier.fillMaxSize()) {
-        val visible = clanList.isNotEmpty()
-        FadeAnimation(visible = visible) {
-            LazyVerticalGrid(
-                state = scrollState,
-                columns = GridCells.Adaptive(getItemWidth())
-            ) {
-                items(
-                    items = clanList,
-                    key = {
-                        it.clanBattleId
-                    }
-                ) {
-                    ClanBattleItem(it, toClanBossInfo)
-                }
-                item {
-                    CommonSpacer()
-                }
-            }
-        }
-        FadeAnimation(visible = !visible) {
-            LazyVerticalGrid(
-                columns = GridCells.Adaptive(getItemWidth())
-            ) {
-                items(20) {
-                    ClanBattleItem(ClanBattleInfo(), toClanBossInfo)
-                }
-            }
-        }
-        //回到顶部
-        FabCompose(
-            iconType = MainIconType.CLAN,
-            text = stringResource(id = R.string.tool_clan),
-            modifier = Modifier
-                .align(Alignment.BottomEnd)
-                .padding(end = Dimen.fabMarginEnd, bottom = Dimen.fabMargin)
-        ) {
-            coroutineScope.launch {
-                try {
-                    scrollState.scrollToItem(0)
-                } catch (e: Exception) {
-                }
-            }
-        }
-    }
-
-}
-
-/**
- * 图标列表
- */
-@Composable
-private fun ClanBattleItem(
-    clanBattleInfo: ClanBattleInfo,
-    toClanBossInfo: (Int, Int, Int) -> Unit,
-) {
-    val placeholder = clanBattleInfo.clanBattleId == -1
-    val bossUnitIdList = clanBattleInfo.unitIds.intArrayList.subList(0, 5)
-
-
-    Column(
-        modifier = Modifier.padding(
-            horizontal = Dimen.largePadding,
-            vertical = Dimen.mediumPadding
-        )
-    ) {
-        //标题
-        Row(
-            modifier = Modifier.padding(bottom = Dimen.mediumPadding),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            //日期
-            MainTitleText(
-                text = clanBattleInfo.getDate(),
-                modifier = Modifier.placeholder(
-                    visible = placeholder,
-                    highlight = PlaceholderHighlight.shimmer()
-                )
-            )
-            //阶段数
-            MainTitleText(
-                text = stringResource(
-                    id = R.string.phase,
-                    getZhNumberText(clanBattleInfo.phase)
-                ),
-                backgroundColor = getSectionTextColor(clanBattleInfo.phase),
-                modifier = Modifier
-                    .padding(start = Dimen.smallPadding)
-                    .placeholder(
-                        visible = placeholder,
-                        highlight = PlaceholderHighlight.shimmer()
-                    ),
-            )
-        }
-
-        MainCard(
-            modifier = Modifier
-                .placeholder(
-                    visible = placeholder,
-                    highlight = PlaceholderHighlight.shimmer()
-                )
-        ) {
-            //图标
-            Row {
-                bossUnitIdList.forEachIndexed { index, it ->
-                    Box(
-                        modifier = Modifier.padding(Dimen.mediumPadding),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        IconCompose(
-                            data = ImageResourceHelper.getInstance().getUrl(ICON_UNIT, it)
-                        ) {
-                            if (!placeholder) {
-                                toClanBossInfo(
-                                    clanBattleInfo.clanBattleId,
-                                    index,
-                                    clanBattleInfo.phase
-                                )
-                            }
-                        }
-                        //多目标提示
-                        val targetCount = clanBattleInfo.getMultiCount(index)
-                        if (targetCount > 0) {
-                            MainTitleText(
-                                text = targetCount.toString(),
-                                modifier = Modifier.align(Alignment.BottomEnd)
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-}
 
 /**
  * 团队战 BOSS 详情
  */
 @OptIn(ExperimentalPagerApi::class)
 @Composable
-fun ClanBossInfoPager(
+fun ClanBattleDetail(
     clanBattleId: Int,
     index: Int,
     maxPhase: Int,
@@ -253,7 +91,7 @@ fun ClanBossInfoPager(
                 clanBattleValue.unitIdList.forEach {
                     urls.add(
                         ImageResourceHelper.getInstance()
-                            .getUrl(ICON_UNIT, it)
+                            .getUrl(ImageResourceHelper.ICON_UNIT, it)
                     )
                 }
                 IconHorizontalPagerIndicator(pagerState = pagerState, urls = urls)
@@ -263,7 +101,7 @@ fun ClanBossInfoPager(
                     state = pagerState,
                 ) { pagerIndex ->
                     if (bossDataList.isNotEmpty()) {
-                        ClanBossInfoPagerItem(
+                        ClanBattleBossItem(
                             pagerIndex,
                             clanBattleValue.getMultiCount(pagerIndex) > 0,
                             bossDataList,
@@ -299,7 +137,7 @@ fun ClanBossInfoPager(
  * Boss 信息详情
  */
 @Composable
-private fun ClanBossInfoPagerItem(
+private fun ClanBattleBossItem(
     pagerIndex: Int,
     isMultiEnemy: Boolean,
     bossDataList: List<EnemyParameterPro>,
@@ -379,6 +217,7 @@ private fun ClanBossInfoPagerItem(
 
 }
 
+
 /**
  * Boss 技能信息
  */
@@ -419,34 +258,6 @@ fun BossSkillList(
                     toSummonDetail = toSummonDetail
                 )
             }
-        }
-    }
-}
-
-
-/**
- * 获取团队战阶段字体颜色
- */
-@Composable
-fun getSectionTextColor(section: Int): Color {
-    val color = when (section) {
-        1 -> R.color.color_rank_2_3
-        2 -> R.color.color_rank_4_6
-        3 -> R.color.color_rank_7_10
-        4 -> R.color.color_rank_11_17
-        else -> R.color.color_rank_18_20
-    }
-    return colorResource(id = color)
-}
-
-@Preview
-@Composable
-private fun ClanBattleItemPreview() {
-    PreviewBox {
-        Column {
-            ClanBattleItem(
-                clanBattleInfo = ClanBattleInfo(1001),
-                toClanBossInfo = { _, _, _ -> })
         }
     }
 }
