+++ app/src/main/java/cn/wthee/pcrtool/ui/tool/pvp/PvpFloatSearch.kt
@@ -32,6 +32,7 @@ fun PvpFloatSearch(spanCount: Int, pvpViewModel: PvpViewModel = hiltViewModel())
     val showResult = navViewModel.showResult.observeAsState().value ?: false
     val pagerState = rememberPagerState()
     val selectListState = rememberLazyGridState()
+    val usedListState = rememberLazyGridState()
     val resultListState = rememberLazyGridState()
     val favoritesListState = rememberLazyGridState()
     val historyListState = rememberLazyGridState()
@@ -89,6 +90,7 @@ fun PvpFloatSearch(spanCount: Int, pvpViewModel: PvpViewModel = hiltViewModel())
                         initSpanCount = spanCount,
                         pagerState = pagerState,
                         selectListState = selectListState,
+                        usedListState = usedListState,
                         resultListState = resultListState,
                         favoritesListState = favoritesListState,
                         historyListState = historyListState,
